fetchcraft-demos/fetchcraft-demo-mcp-server/pyproject.toml
[project]
name = "fetchcraft-demo-mcp-server"
version = "0.4.1"
description = "Demo MCP server for Fetchcraft"
requires-python = ">=3.12,<3.14"
dependencies = [
  "fetchcraft-core",
  "pydantic-settings>=2.11.0",
]

[project.scripts]
fetchcraft-mcp-server = "fetchcraft.demo.mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
sources = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/fetchcraft/"]

[tool.uv]
src-layout = true

[tool.uv.sources]
fetchcraft-core = { workspace = true }
